pitch control on one osc, and fixed sluiggish fifo sm buffer

+++ arduino_libmyriad/myriad_messages.h
@@ -8,5 +8,5 @@ size_t maxOscBankType = 1;
 
 struct spiMessage {
     uint8_t msg;
-    size_t value;
+    float value;
 };
